+++ crates/nostr/CHANGELOG.md
@@ -31,6 +31,7 @@
 
 ### Changed
 
+- Rename `NostrConnectURI` to `NostrConnectUri`
 - Rename `NostrWalletConnectURI` to `NostrWalletConnectUri`
 - Make `Kind::from_u16` and `Kind::as_u16` const (https://github.com/rust-nostr/nostr/pull/1136)